Fix of issue 105 ready for merge (#111)
* Removed most codeScene warnings

* All codescene warnings fixed
