+++ routes/studentRoutes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const errorCodes = require('../helpers/errorCodes');
+const { markAsCompleted } = require('../helpers/answerExercises');
 const requireLogin = require('../middlewares/requireLogin');
 const mongoose = require('mongoose');
 const { encrypt, compare } = require('../helpers/password');
@@ -123,7 +124,11 @@ async function updateUserLevel(userId, points, level) {
   }
 
   // Update user points and level in the database
-  await StudentModel.findOneAndUpdate({ baseUser: userId }, { points: points, level: level });
+  await StudentModel.findOneAndUpdate(
+    { baseUser: userId },
+    { $inc: { points: points }, $set: { level: level } },
+    { new: true } // Set to true if you want to get the updated document as a result
+  );
 }
 
 // Mark courses, sections, and exercises as completed for a user
@@ -164,191 +169,5 @@ router.patch('/:id/completed', requireLogin, async (req, res) => {
     });
   }
 });
-
-async function markAsCompleted(user, exerciseId, points, isComplete) {
-
-  // Retrieve the exercise by ID to find sectionId
-  const exercise = await ExerciseModel.findById(exerciseId);
-
-  if (!exercise) {
-    throw errorCodes['E0012'];
-  }
-
-  const sectionIdString = exercise.parentSection;
-  const sectionId = mongoose.Types.ObjectId(sectionIdString.toString());
-  const section = await SectionModel.findById(sectionId);
-
-  if (!section) {
-    throw errorCodes['E0008'];
-  }
-
-  const courseIdString = section.parentCourse;
-  const courseId = mongoose.Types.ObjectId(courseIdString.toString());
-
-
-  await markExerciseAsCompleted(user, courseId, sectionId, exerciseId, points, isComplete);
-
-  // Check if all exercises in the section are completed
-  user = await StudentModel.findOne({ baseUser: user.baseUser });
-
-  // Redefine the completedCourseIndex with the updated user
-  const completedCourseIndex = user.completedCourses.findIndex(completedCourse => completedCourse.courseId.equals(courseId));
-  const completedSection = user.completedCourses[completedCourseIndex].completedSections.find(completedSection => completedSection.sectionId.equals(sectionId));
-
-  // Check if all exercises are present and marked as complete
-  const allExercisesCompleted = completedSection.completedExercises.length === section.exercises.length &&
-    completedSection.completedExercises.every(completedExercise => completedExercise.isComplete);
-
-
-
-  // Update section's isComplete status, and afterwards update the points for the specific course and section for the user.
-  await StudentModel.findOneAndUpdate(
-    { baseUser: user.baseUser },
-    {
-      $set: {
-        [`completedCourses.${completedCourseIndex}.completedSections.$[section].isComplete`]: allExercisesCompleted,
-        [`completedCourses.${completedCourseIndex}.completedSections.$[section].completionDate`]: Date.now()
-      }
-    },
-    {
-      arrayFilters: [{ 'section.sectionId': sectionId }]
-    }
-  ).post(async function() {
-    // Calculate the totalPoints for the updated section
-    const completedCourse = user.completedCourses[completedCourseIndex];
-    const updatedSection = completedCourse.completedSections.find(completedSection => completedSection.sectionId.equals(sectionId))
-
-    updatedSection.totalPoints = updatedSection.completedExercises.reduce((total, exercise) => total + exercise.pointsGiven, 0);
-
-    // Update the totalPoints for the completedCourse
-    completedCourse.totalPoints = completedCourse.completedSections.reduce((total, section) => total + section.totalPoints, 0);
-
-    // Update the user's totalPoints
-    await StudentModel.findOneAndUpdate(
-      { baseUser: user.baseUser },
-      {
-        $set: {
-          [`completedCourses.${completedCourseIndex}.totalPoints`]: completedCourse.totalPoints,
-          [`completedCourses.${completedCourseIndex}.completedSections.$[section].totalPoints`]: updatedSection.totalPoints
-        }
-      },
-      {
-        arrayFilters: [{ 'section.sectionId': sectionId }]
-      }
-    );
-  });
-
-  user = await StudentModel.findOne({ baseUser: user.baseUser });
-
-  // Check if all sections in the course are completed
-  const allSectionsCompleted = user.completedCourses[completedCourseIndex].completedSections.every(completedSection => completedSection.isComplete);
-
-  // Update course's isComplete status
-  await StudentModel.findOneAndUpdate(
-    { baseUser: user.baseUser },
-    {
-      $set: {
-        [`completedCourses.${completedCourseIndex}.isComplete`]: allSectionsCompleted,
-        [`completedCourses.${completedCourseIndex}.completionDate`]: Date.now()
-      }
-    }
-  );  
-
-  return await StudentModel.findOne({ baseUser: user.baseUser });
-}
-
-async function markExerciseAsCompleted(user, courseId, sectionId, exerciseId, points, isComplete) {
-  const completedCourseIndex = user.completedCourses.findIndex(completedCourse => completedCourse.courseId.equals(courseId));
-  if (completedCourseIndex === -1) {
-    // Course not found, add it along with completedSections and completedExercises
-    await StudentModel.findOneAndUpdate(
-      { baseUser: user.baseUser },
-      {
-        $push: {
-          completedCourses: {
-            courseId,
-            completedSections: [{ sectionId, completedExercises: [{ exerciseId, isComplete, pointsGiven: points }] }]
-          }
-        }
-      }
-    );    
-  } else {
-    const completedSectionIndex = user.completedCourses[completedCourseIndex].completedSections.findIndex(completedSection => completedSection.sectionId.equals(sectionId));
-
-    if (completedSectionIndex === -1) {
-      // Section not found, add it along with completedExercises
-      await StudentModel.findOneAndUpdate(
-        { baseUser: user.baseUser },
-        {
-          $push: {
-            'completedCourses.$.completedSections': {
-              sectionId,
-              completedExercises: [{ exerciseId, isComplete, pointsGiven: points }]
-            }
-          }
-        }
-      );      
-    } else {
-      let exerciseFound;
-
-      // Check if the exercise is in the user's completedExercises array
-      user.completedCourses[completedCourseIndex].completedSections.some(section => {
-        section.completedExercises.forEach(exercise => {
-          if (exercise.exerciseId == exerciseId) {
-            // Found the matching exerciseId, set exerciseIsComplete to the associated isComplete value
-            exerciseFound = exercise;
-          }
-        });
-      });
-
-
-      // If the exercise is already marked as complete, check it's isComplete value, if true throw error,
-      // else update the exercise field "isComplete" to true, and points to the points given (5)
-      if (exerciseFound) {
-        if (exerciseFound.isComplete) {
-          throw errorCodes['E0801'];
-        } else {
-          // Update the exercise field "isComplete" to true, and points to the points given
-          await StudentModel.updateOne(
-            {
-              baseUser: user.baseUser,
-              'completedCourses.courseId': courseId,
-              'completedCourses.completedSections.sectionId': sectionId,
-              'completedCourses.completedSections.completedExercises.exerciseId': exerciseId
-            },
-            {
-              $set: {
-                'completedCourses.$.completedSections.$[section].completedExercises.$[exercise].pointsGiven': points,
-                'completedCourses.$.completedSections.$[section].completedExercises.$[exercise].isComplete': isComplete,
-                'completedCourses.$.completedSections.$[section].completedExercises.$[exercise].completionDate': Date.now()
-              }
-            },
-            {
-              arrayFilters: [{ 'section.sectionId': sectionId }, { 'exercise.exerciseId': exerciseId }]
-            }
-          );
-          return;
-        }
-      }
-
-      // Adds the exercise to the user's completedExercises array
-      await StudentModel.findOneAndUpdate(
-        {
-          baseUser: user.baseUser,
-          'completedCourses.courseId': courseId,
-          'completedCourses.completedSections.sectionId': sectionId
-        },
-        {
-          $addToSet: {
-            'completedCourses.$.completedSections.$[section].completedExercises': { exerciseId, isComplete, pointsGiven: points }
-          }
-        },
-        {
-          arrayFilters: [{ 'section.sectionId': sectionId }]
-        }
-      );      
-    }
-  }
-}
 
 module.exports = router;
